[BLE] Implementation of Bluetooth manager complete and tested!
